split training and making transfer values into 2 scripts
